Cabal2Nix.hs: simplified the command line parser logic
I have examined two libraries that are supposed to manage command line
flags, cmdlib and cmdargs, and both of them are unsatisfactory. The
cmdargs library seems to work well, but it's wildly impure, i.e. the
parser will function cannot be re-run in ghci! Also, it automatically
adds flags to the command line (such as --verbose) and stores their
state in some magic, global variable. It may work well in practice, but
as a matter of principle I just don't want that kind of thing in my
Haskell programs.

The other library, cmdlib, seems to be better in that regard, but it's
API is somewhat incomplete. It can configure wildly complex interfaces
(such as the one from darcs), but *simple* interfaces don't seem to
work well, i.e. there is no way to write a program that doesn't use
sub-commands. Furthermore, I managed to crash the code three or four or
times, which doesn't inspire trust either.

So for the time being, we seem to be stuck with getopts.