Cabal2Nix/PostProcess.hs: refer to libc as 'libc' rather than 'glibc'
Some packages need libc to include headers like <time.h>, etc. To satisfy
that dependency, we add a reference to the non-existent attribute 'libc',
which ought to be overridden as follows in haskell-packages.nix:

  foo = callPackage ../development/libraries/haskell/foo {
    libc = pkgs.stdenv.gcc.libc;
  };

On impure platforms, such as Darwin, that reference will be 'null', but on
Linux it will refer to glibc, as expected.