Cabal2Nix/Generate.hs: configure Cabal to assume we're building with GHC 7.2.2
For the purposes of Nixpkgs, it's usually unimportant which GHC version is
building the package. The Cabal instructions may customize a compiler flag here
and there at build time, but those changes don't affect the Nix expression that
cabal2nix generates.

With the release of aeson 0.4.0.0, however, that has changed, because aeson
needs additional dependencies when compiled with GHC 7.2.2 than it would need
when compiled with, say, GHC 7.0.4. Now, that is a classical choose-your-poison
situation. We cannot generate on expression that is "correct" for both
compilers, still we would like to be able to compile aeson with both GHC 7.0.4
and 7.2.2. To work around that problem, we over-specify the compiler version,
because the 7.2.2 version happens to compile fine with 7.0.4, but not the other
way round. The resulting expression is inaccurate, but it works.

That kludge is going to fail, of course, if we want to support a package where
we would have to under-specify the compiler version to make it build. Let's
hope it just won't happen.