src/Cabal2Nix/CorePackages.hs: dropped 'filepath' and 'Cabal' from that list
These libraries *are* core packages, but still we would like to be able to
treat them like any other Haskell library in Nix, i.e. we would like to
override the version used of any of those libraries during a build if the
version that comes with GHC is too old.