enable GIO support in GTK package
The flag enable-gio is True by default, so I suppose we should also
build GTK with GIO in nixpkgs. The hackage haddocks are also built with
GIO enabled, so it can be confusing if GIO is disabled in nixpkgs' gtk
version and thus some functions that are documented on hackage are
missing.