hackage4nix: disable Hydra builds for all "old" Haskell packages
Old means that we have a newer version in Nixpkgs already. If there is a reason
for the old package to still exist, i.e. because some other package is still
referring to it, then Hydra will still build the package as a dependency, but
we won't want to build old packages on their own right.