don't try different fetcher if cabal parsing fails
If we get a directory which contains a cabal file that produces a parse
error, we don't try another fetcher. Trying another fetcher when parsing
fails is only needed if the fetcher result is a file, because then it
might be that the fetcher only got a html result (fetchurl with git repo
for example) which doesn't parse (=> need to try another fetcher).