Fix generation from local archives and rel. paths
`nix-prefetch-zip` cannot handle relative paths. Therefore,
we need to resolve the relative path before passing it to
`nix-prefetch-zip`. You also probably don't want a relative path in your
derivation when generating from an archive file.

Also, `cabal2nix` now automatically adds a `./` in front of paths that
contain no slash (this is needed for nix to recognize the path).