hackage2nix: filter the package name from the list of missing libraries
Sometimes, packages refer to their own library in executables or test
suites with an incorrect version bound. In those cases, the library
currently being built ends up in the list of missing dependencies. We
cannot generate a 'null' override for it, however, because the generate
expression doesn't take that name as an argument.