Implement a wacky work-around for https://github.com/NixOS/cabal2nix/issues/136.
The post-process hook filters any package from pkgconfigDepends that is
also part of buildDepends. In other words, if some package depends on
the Haskell library "glib", then we filter "glib" from pkgconfigDepends.
As a result, the bogus resolver in "hackage2nix" won't fire.

Of course, this means that the system library is now missing from those
builds that had their pkgconfigDepends stripped. We counter this by
adding the appropriate system package to the propagatedBuildInputs field
of the their corresponding Haskell library. For example, the Haskell
package "glib" exports the system library "glib" to all its users -- so
these users don't have to depend on it anymore.

The real solution is to fix the resolver, of course, but this change
will hopefully fix a number of builds *right now* without much of an
effort to do the right thing(tm).