Improve memory usage of package set generation
We really want to avoid memorizing the GenericPackageDescriptions, since
those tend to be quite big. The commit achieves this by not storing the
GenericPackageDescriptions in a Map, but instead only storing IO actions
that produce them. This way, package descriptions are never cached.

To avoid building up a huge list in writePackageSet, we also only apply
concurrency in batches of up to 100 packages. That way, we still get enough
concurrency, but we don't have to wait for all of the 6000 packages to be
generated before we can start writing out results.