PostProcess.hs: extend the fixGtkBuilds function to re-produce the old state of hackage-packages.nix
With the new data structures in place, we can generate a more accurate version
of hackage-packages.nix. I'd like to tackle this incrementally, though, so I've
added a couple of nasty hacks to the post-processor that don't make much sense
but ensure that the hackage-packages.nix file we generate looks pretty much
like the previous one.