nix-paths: fix dist target
The auto-generated source file worked fine, but cabal failed the sdist
target because it can't deal with a package that doesn't contain any
source code files. To remedy that issue, we now *have* a source code
file and inject the actual paths from Setup.hs via CPP, similarly to
ghc-paths does it.