hmatrix-special: drop hard-coded /usr path from cabal file
The prior code did a sed substitution of the dangerous /usr/ path in
hmatrix.cabal with a safer /dont/hardcode/path one. That's fine for
compilation of hmatrix, but hmatrix-special (which depends on hmatrix)
then fails to compile because an impure (outside /nix/store) link is
detected. So this commit deletes the problematic reference in
hmatrix.cabal, rather than substitute it.

Fixes https://github.com/NixOS/cabal2nix/issues/307.
Closes https://github.com/NixOS/cabal2nix/pull/308.