hmatrix: disable "sed" hack for versions newer than 0.18.1.0.
hmatrix hard-codes all kinds of bogus paths into the Cabal file which break the
build with Nix. https://github.com/albertoruiz/hmatrix/pull/219 fixed this
issue, but unfortunately a version that contains these changes was never
released (https://github.com/albertoruiz/hmatrix/issues/243). This means that
we still need the "sed" hack in place for all current releases of hmatrix, but
applying that hack to a current git version of hmatrix breaks the build.

To work around this issue, we restrict the application of the sed hack to
versions older than 0.18.1.1, which is what current git version call
themselves. Therefore, a call like

  $ cabal2nix -fdisable-default-paths --subpath=packages/base https://github.com/albertoruiz/hmatrix

generates a Nix file without the "sed" hack, but Nix builds for the current
release versions still have it.

Fixes https://github.com/NixOS/cabal2nix/issues/327.