normalizeCabalFlags: fix broken precedence logic
The function assumed that given a sequence of multiple instances of the same
flag, Cabal chooses the value given first, but in fact it chooses the value
given last. Not sure whether this behavior ever changed at some point in time,
but it's certainly more in line with what one would expect.