finalizeGenericPackageDescription: don't rely on defaultComponentRequestedSpec
A record update on defaultComponentRequestedSpec is no safer than just
constructing an instance of the type directly, because there is a mysterious
second constructor, OneComponentRequestedSpec, that our record update won't
match if defaultComponentRequestedSpec turns out to be such a thing for some
reason. Curiously enough, we have helper functions that can query the state of
the ComponentRequestedSpec type, but we cannot easily manipulate it.