If both package.yaml and cabal exist, prefer package.yaml.
This is because in a nomal workflow, you would generate the .cabal
every now and then from the package.yaml file.

If you forget to do this generation and you have an old .cabal file
in your tree (common because for users of hpack, .cabal files are
usually in .gitignore), then until now nix would pick up the outdated
cabal file and you would accidentally be running against outdated cabal
contents.

So in this commit we change the logic from

  "if *.cabal doens't exist, check if package.yaml exist, if so use hpack"

to

  "if package.ymal exists, use hpack, otherwise use *.cabal"