Fail with an error if the set does not contain the binding we're trying to replace.
The 'replace' function is used to fix bindings in 'Derivation' that the
'resolveInNixpkgs' lookup failed to figure out. Now, one problem we encountered
is that the (bogus) binding we assumed to be there actually doesn't exist any
more because Nixpkgs has changed, and in such a case 'replace' silently did
nothing, resulting in a potentially broken build.

Instead, we interpret "replace old new bs" to imply that "old `elem` bs" is
true and fail if it's not to indicate that the post-process code must be
adapted to the new situation.