Normalize flags to all-lowercase before handing them to Cabal.
Cabal expects the payload of the "opaque" FlagName type to be spelled in all
lower-case characters. The mkFlagName function, however, does not ensure that
invariant. We need a good solution to avoid this pitfall. For the time being,
we simply convert all flags used in this module to lower-case ourselves, but
chances are that there are other places in cabal2nix that create improper
FlagName instances through the official API.

Related: https://github.com/NixOS/nixpkgs/issues/73392