fix fetching from remote archives
Commit 854695a removed attempting to use nix-prefetch-zip on the
downloaded artifact in favor of only nix-prefetch-url, which does not
unpack automatically. With that, fetching from remote archives became
broken as no handler would succeed.
This adds an optional `command` argument to fetchWith, with which the
command to use for fetching can be specified instead of it being derived
from the kind.
That `command` argument is then used by the reintroduced kind `zip` to
call nix-prefetch-url (with the `--unpack` flag) instead of
nix-prefetch-zip.
In addition, a bug where `fetchWith` would return the path of the
artifact instead of the output of the handler has been fixed.