Generate predictable results for "description".
Some packages (BioHMM-1.2.0) have trailing white-space at the end of
their synopsis. This breaks our description normalizing function, which
checks whether the last character of the description is a dot -- and
strips it if it is. Now, newer versions of Cabal actually strip that
trailing white-space, but since we want to be able to compile against
older Cabals, too, we can't rely on Cabal to do that for us. We'll
rather strip all trailing white-space ourselves.

Fixes https://github.com/NixOS/cabal2nix/issues/459.
Closes https://github.com/NixOS/cabal2nix/pull/460.