Update hackage2nix instructions to reflect cabal v2 usage
Seems like this README was not updated when stack was removed from this
repository and the old style cabal workflow was dropped in favor of
cabal v2-*. Now we need to build cabal2nix using cabal v2-build in order
for the cabal v2-run in update-nixpkgs.sh to work.

Also add a note about required software for building and running hackage2nix.