FromCabal.Name: resolve library lzma to xz
lzma is only an alias to xz in nixpkgs nowadays which leads to all kinds
of evaluation troubles (since such aliases are stripped and not present
if aliases are disabled).

To fix this, we preemptively respect the alias in
Distribution.Nixpkgs.Haskell.FromCabal.Name.

Reference https://github.com/NixOS/nixpkgs/issues/118996.