Distribution.Nixpkgs.PackageMap: collect all attribute paths
`nix-env` evaluates nixpkgs and prints a collection of *derivations*
with additional meta data (like name, attribute path, description, …).
Derivations are uniquely identified via their `outPath`, so `nix-env`
will by design only return one entry *per derivation*. This is
undesireable in the case of `resolve` where we try to match on the
*attribute name*. `nix-env` fails us here if there are multiple
attribute names for a single derivation: In such cases it only prints
the one that comes first alphabetically.

For example when resolving `pam`, `resolve` would previously return
`python38Packages.pam` despite obviously just `pam` would be correct.
This is caused by `pam` and `linux-pam` evaluating to the same
derivation on Linux and `nix-env` printing only `linux-pam`
consequently.

To fix this we employ a relatively simple nix expression which takes a
path to nixpkgs and optional arguments to pass to it and returns a list
of all attribute paths which point to a derivation. In doing that it
also respects the rules surrounding `recurseForDerivations`.

This also greatly simplifies the parsing code since we no longer need to
split and parse the attribute paths. Also there is a significant speedup
of ~125% (on my laptop) since we don't need to compute any `outPath`s.

However, we are forced to use `nix-instantiate` instead of `nix-env` for
this which supports different flags. Thus we can't really keep the
`extraArgs` approach we had previously. `readNixpkgPackageMap` now
expects a (nix) path to nixpkgs and optionally a attribute set to pass
to nixpkgs as an argument. The version number has been increased to
1.6.0 as a result.

Resolves #9.