parsePackageData: fix preferred version ranges parsing
Previously we used `Dependency` for this which became impossible with
Cabal 3.4.0.0 since the third field became a `NonEmptySet` in that
version whereas in our use case it would always be empty.

We can't however parse the preferred version ranges file in an isolated
fashion, since it always also contains the package name. This leads to a
parse failure for all packages that have the preferred-version-ranges
file, like postgresql-simple (see https://github.com/NixOS/cabal2nix/issues/501).

The solution is to use `PackageVersionConstraint` which seems to be the intended
type to use instead of `Dependency` for these kinds of applications.