Don't use hpack if an appopriate cabal file is found
hpack will only be used if no appropriate cabal file is found in the
source. This now means that cabal2nix will see the cabal file as the
single source of truth and prevent the previous problem that hpack would
fail when building if the checked-in cabal file was generated with a
different hpack version than is in nixpkgs / used by the builder.