README: suggest shared development env for cabal2nix related packages
This is basically what I've used so far and works quite well without
being too opinionated and using mostly vanilla cabal.

Also gets rid of the outdated suggestion to use stack which isn't
possible anymore.

I'm not sure this is the ideal way we should settle on, but it is a
quite comfortable solution that doesn't require any changes at this
point.