cabal2nix.cabal: update maintainer, only list original author
The list of contributors hasn't been updated in over 3 years now, so
let's just fill the author field with what the cabal manual prescribes:
The original author.

I'm happy to respond to any email that comes in via the maintainer email,
this field can also always be updated via the hackage revision mechanism.