Distribution.Nixpkgs.Haskell.OrphanInstances: parse `aarch64-darwin`
To support `aarch64-darwin` in the `unsupported-platforms`
configuration, we need to correctly parse it in the `FromJSON` instance
of `Platform`.

We increase the lower bound on `distribution-nixpkgs` in order to
prevent users from being able to parse a configuration which will
result in `hackage2nix` crashing while rendering the resulting nix
file (`aarch64-darwin` support was added to distribution-nixpkgs in
version 1.6.1).