cabal2nix: Assume "unknown" hackage license to be free
We are currently very inconsequent in regard to "unknown" licenses. In
general they stem from packages having their license field set to
"LicenseRef-OtherLicense" accompanied by a LICENSE file which specifies
the real license.

Currently in nixpkgs "unknown" licenses result in hydra builds being
disabled for that package, but the package is not treated as "unfree" by
anything else in nixpkgs. Users don‘t get warnings when they use them
and we actually very often build those packages on hydra because they
are dependencies on other packages with free licenses.

This change simply acknowledges the fact that we are already building
and distributing most of these packages by giving them their own hydra
job.