Distribution.Nixpkgs.Haskell.FromCabal: don't depend on webkitgtk24x
webkitgtk24x-gtk{2,3} have been removed from nixpkgs a while ago and
only continue to exist as aliases which throw, informing about the
removal. This is informative, but depends on evaluating with aliases
enabled which may not always be the case.

Instead we just remove this now essentially useless code from
cabal2nix. This will usually cause a dependency on the current version
of webkitgtk and the build to fail -- then we can mark the resulting
package as broken normally which was prevented before by eval
oddities (I presume at least).