feat(Meta): add Pretty instance for NixpkgsPlatform
Expose fromCabalPlatform (renamed to nixpkgsPlatformFromCabal) and
platformPath to the user via the Pretty instance for NixpkgsPlatform and
also use it internally for rendering. This has some nice side effects:

* We can test the platform rendering in the hspec test suite. This is
  particularly important now, since Cabal may change the way it renders
  platforms in a way that is no longer compatible with nixpkgs at some
  point.

* We can drop partitionPlatforms since we no longer need to obtain the
  inner types -- instead we can just rely on plain Set.partition.

Note that this commit affects the ordering of rendered platform meta
attributes, since we now order the `Platform`s instead of the rendered
Nix strings.