Allow NixpkgsPlatformGroup, use distribution-nixpkgs for parsing
This change no longer exclusively produces `NixpkgsPlatformSinge`s. The
main hurdle for this was parsing: We need to be able to denote platform
groups in the yaml configuration somehow. nixpkgsPlatformFromString
which was added in <https://github.com/NixOS/distribution-nixpkgs/pull/19>
fills in this gap: Any string prefixed with either `platforms.` or
`lib.platforms.` will be interpreted as a platform group, all other
strings will (roughly) be treated as before.