Distribution.Nixpkgs.Haskell.Platform: unify CLI platform parsing
In the cabal2nix project there are two flags that accept some kind of
string that needs to be converted to a `Platform` which is used for
evaluating a `.cabal` file:

* cabal2nix: `--system` accepts `cpu-vendor-os` and `cpu-os`
* hackage2nix: `--platform` only accepts `cpu-os`.

Their names should be swapped, ironically compared to what they do which
should be tackled eventually.

Initially I thought the parsing of these could be replaced by
distribution-nixpkgs new capabilities, but ran into the following issue:
distribution-nixpkgs needs to be lossless, since `Platform`s need to be
rendered again.

This is problematic when evaluating Cabal files, however: Cabal has no
notion of powerpc64le and powerpc64 and GHC's conversion logic causes
them to be both interpreted as PPC64. Package authors likely expect this
behavior, so we risk accidentally getting wrong results by insisting on
losslessness.

To tackle this, we'll have two different ways to parse a
nixpkgs-originating string into a `Platform`:

* One in distribution-nixpkgs to parse into a `NixpkgsPlatformSingle`
* One in cabal2nix to parse into a genuine Cabal `Platform`