hackage2nix: don't discard platforms meta attr from postprocessing
Distribution.Nixpkgs.Haskell.FromCabal.PostProcess defines logic that in
some cases prescribes a list of platforms. Currently this gets discarded
in hackage2nix which is a bug introduced in #506.

Before said PR, hackage2nix used a fairly elaborate logic involving Set
subtraction which was necessary because the default value would be
allKnownPlatforms. Instead of that we can now use a very simple logic:
We use whatever is set or not set by PostProcess / cabal2nix in general
and then allow whatever is yielded by that to be overwritten by the
configuration passed to hackage2nix.