distribution-nixpkgs: accomodate changed platform representations
Cabal 3.8 introduced the following new platform related type
constructors:

* S390X, the 64 bit Arch variant of S390
* Wasm32, the 32 bit WASM Arch
* Wasi, the WebAssembly System Interface, a new “OS”

This means that some of the tuples we were checking have a changed
representation when parsed and we need to update the test suite.

We don't need to adjust cabal2nix nor the related packages anywhere
else, as we don't hardcode the old representations for these OSes /
Arches anywhere.