hackage-db: support cabal-install 3.10.1.0 XDG paths.
From 3.10.1.0 onwards, cabal-install uses XDG directories to store its
state and cache components.  cabal2nix expects to find the Hackage
tarball in ~/.cabal, where it may no longer be located.

This commit changes cabal2nix to look in ~/.cabal if that directory
exists (which matches the behaviour of cabal-install itself), but
otherwise looks in the appropriate XDG directory.