hackage-db: Distribution.Hackage.DB.Path: support CABAL_DIR env var
Just like cabal-install, if CABAL_DIR is set, use that for all
operations. The environment variable does of course not support any
split between $XDG_CONFIG_HOME/cabal and $XDG_CACHE_HOME/cabal which
seems like an oversight on upstream's part, but we don't need to care:
we only ever care about the state dir, so users can just need to set
that.