Setup tz package overrides to reference system tzdata package correctly
This commit changes the cabal2nix hooks for the `tz` Haskell package.
Originally, the `tz` package had a hook that added a phase override for
`preConfigure` that looked like the following:

```nix
preConfigure = "export TZDIR=${pkgs.tzdata}/share/zoneinfo";
```

This works for `hackage2nix`, but the problem is that when using
`cabal2nix` by itself, this produces a Nix file like the following:

```nix
{ mkDerivation, base, binary, bytestring, containers, criterion
, data-default, deepseq, HUnit, lens, lib, QuickCheck, tasty
, tasty-hunit, tasty-quickcheck, tasty-th, template-haskell, thyme
, time, timezone-olson, timezone-series, tzdata, vector
}:
mkDerivation {
  pname = "tz";
  version = "0.1.3.6";
  preConfigure = "export TZDIR=${pkgs.tzdata}/share/zoneinfo";
  ...
}
```

This is trying to reference `pkgs.tzdata`, but `pkgs` is not available
here!

This commit changes the hook for `tz` to take a `system-tzdata` package
as an argument, and then correctly reference this `system-tzdata` for
the tests:

```nix
{ mkDerivation, base, binary, bytestring, containers, criterion
, data-default, deepseq, HUnit, lens, lib, QuickCheck
, system-tzdata, tasty, tasty-hunit, tasty-quickcheck, tasty-th
, template-haskell, thyme, time, timezone-olson, timezone-series
, tzdata, vector
}:
mkDerivation {
  pname = "tz";
  version = "0.1.3.6";
  preCheck = "export TZDIR=${system-tzdata}/share/zoneinfo";
  ...
}
```