cabal2nix: don't emit man postInstall script for pandoc >= 3.1.10
The man page was moved from the pandoc release tarball to the pandoc-cli
tar ball. We'll just stop emitting this postInstall portion as soon as
the pandoc release tar ball no longer has the man page files. Instead
we'll just manually have this code in nixpkgs as an override, so we are
slowly working towards #504.