FromCabal.Name: add missing (versioned) webkitgtk pkg-config names
When we don't know the exact version from the cabal file, we still emit
the generic name that has recently been removed from nixpkgs safe for an
alias (https://github.com/NixOS/nixpkgs/pull/345611) which we can't use
in hackage-packages.nix. This is still the best solution for downstream
expressions generated by cabal2nix, in nixpkgs we'll have to manually
specify the correct version (instead of hackage2nix picking the default
which may not always be correct).

Co-authored-by: sternenseemann <sternenseemann@systemli.org>