shell.nix: add minimal flag
The minimal flag is intended to disable “superfluous” stuff, i.e.
anything that isn't necessary to build the included packages and run
their tests. It is used to determine withHls for now, but has a less
specific meaning.

The idea is that you use

  nix-shell --arg minimal true --arg pkgs "(import <nixpkgs> {})"

when using a nixpkgs version different from the pin.