FromCabal: remove LLVM evil
Usually, there is no reason to tie use of LLVM libraries to the LLVM
version used by GHC’s LLVM backend as packages should not declare this
as a library dependency.

Also, `libclang` is the correct dependency for Clang used as a
library, rather than the wrapped `clang` tool. (In practice, though,
this doesn’t really help: the packages this changes in Nixpkgs
are already broken because they depend on ancient LLVMs, want the
`llvm-config(1)` tool that is in the `llvm` package, and in one of
the two cases the `Setup.hs` doesn’t even build with modern Cabal.)

To get a consistent result in hackage-packages.nix, we unfortunately
have to add an extra alias for clang-3.8 (which has long been removed)
until LibClang is removed from the package set.