hackage2nix: use Identifier for bindings in hackage-packages.nix
This increases correctness as toNixName is already used to generate the
package expression arguments. If toNixName and attr were to diverge,
packages would fail to resolve their dependencies correctly.

A side effect of this change is that `pPrint attr` only quotes
attribute names as needed, resulting in a huge diff for
hackage-packages.nix.