FromCabal.Name: prefix pkg names with _ if they can't be identifiers
While we can quote attribute names, we can't use quoting to use
arbitrary strings as function arguments in Nix unfortunately. This means
we couldn't previously generate expressions for packages that depend on
packages that start with numbers or clash with keywords.

To fix this, we simply prefix the offending packages with an underscore.
This mirrors an existing Nixpkgs convention (see e.g. pkgs._2bwm). This
mapping work for all Hackage packages and is easily reversible by just
removing the underscore. We can never introduce ambiguity this way since
Cabal doesn't allow underscores in package names.

Cabal packages not conforming to Hackage's restrictions won't work yet.
For this, we would need to implement some kind of encoding.

Resolves #163.
Resolves #164.