hackage2nix: do not generate an expression for acme-everything
The package expression for acme-everything is huge since it has to
list all packages released to Hackage as of 2018 twice. As this package
isn't actually useful and probably doesn't even have a valid install
plan, we can remove it to keep the size of hackage-packages.nix in
check.

In the future, such exclusions should be decided by configurable
exclusions as proposed in https://github.com/NixOS/cabal2nix/pull/667.
It may be a prudent idea to use the spam and acme categories of Hackage
as a baseline for package exclusions.

This partially reverts commit d105bfca2fdacaf7b4153f518f9c1498a7f16581.