Language.Nix.Identifier: correctly implement Nix escape sequences
This code previously assumed escape sequences in strings work like in
Haskell. This is not generally the case. This commit adds a custom
implementation of the relevant parsing and rendering logic which even
simplifies some things.

Escape sequences are a little weird in Nix since only very few things
_have_ to be escaped and some things not always, depending on
context (i.e. $). I've linked the best reference (the lexer) in the
code, but the manual is also helpful:
https://nix.dev/manual/nix/2.30/language/string-literals.html