shell.nix: haskell-ci: 2025-03-30 -> 2025-06-04
Later versions support GHC 9.10.3, but are very annoying to get to work
unfortunately because they seem to depend on a Cabal version that hasn't
been released with any version of GHC yet which would require
maintaining a lot of overrides here.

I think it's fine to just switch to that when 9.12.3 has been released.